fix(fx-core): keep template validation independent of API builds
Move typed package parsing into a pure validation module with caller-owned error construction. Preserve runtime parsing exports and SystemError adaptation, and remove the build-time archive validator's runtime import of packageDir.

Add AC-29 regression coverage that validates source archives and prepares packages in a fresh process where product API runtime imports are forbidden.

Validation: 5840 full fx-core tests passed (93.39% line coverage); final affected rerun 238/238 passed; 29-template build, source and scenario-tooling TypeScript checks, formatting and lint (0 errors) passed. Full npm run setup was not rerun because the execution policy blocked dependency installation; this limitation remains pending CI validation.
